StringToCharacterOrStringConverter implements TemplatedContext was ConverterStringToCharacterOrString extends Converter2
